Fix nicknames being wiped out by sleeper
If there were changes in addition to a nickname then the nickname
would be deleted. This fixes it so that nicknames are only ever
updated if they are not empty.

Save yahoo ids back to DB if they are found when generating power
rankings. Also ensures that yahoo ids aren't removed when players
are updated from sleeper

Refactor the DB code to pull player functions into a separate file
to make the code easier to navigate.

Testing:
- Unit tests are updated and pass
- Limited local testing